requirements.txt
discord.py[voice]
python-dotenv
pdf2image
pillow
